rename: FriendshipBuilder 패키지 이동
